+++ xbird-starter-resteasy/src/main/java/com/github/zhycn/resteasy/boot/JaxrsApplicationScanner.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.zhycn.dubbo.resteasy.boot;
+package com.github.zhycn.resteasy.boot;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;